Optimization and proper gitignore of tmp

+++ intensify.rb
@@ -10,17 +10,23 @@ get '/' do
   content_type 'image/gif'
 
   return "Provide a valid image URL" unless params['url']
+
   image = MiniMagick::Image.open(params['url'])
+  shake = params['shake']&.to_i || 5
 
   # set up paths
-  image_path = "tmp/#{image.path.split("/").last}"
-  intensified_path = "tmp/#{SecureRandom.hex(8)}.gif"
+  image_path = "tmp/#{image.signature}_#{shake}"
+  intensified_path = "tmp/#{image.signature}_#{shake}_intensified.gif"
+
+  # if file already in tmp, return that
+  if File.exists?(intensified_path)
+    puts "CACHED!"
+    return MiniMagick::Image.open(intensified_path).to_blob
+  end
 
   # set up files
   image.write(image_path)
   File.open(intensified_path, "w") {}
-
-  shake = params['shake']&.to_i || 5
 
   puts "Converting..."
